pyproject.toml
[tool.poetry]
name = "logs-python"
version = "0.1.0"
description = ""
authors = ["Arvind <pingarvindforquestions@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
elasticsearch = "6.3.1"
fastapi = "^0.105.0"
pydantic-settings = "^2.1.0"
uvicorn = "^0.24.0.post1"
python-decouple = "^3.8"
gunicorn = "^21.2.0"


[tool.poetry.group.dev.dependencies]
djlint = "^1.34.0"
black = "^23.12.0"
pylint = "^3.0.3"
isort = "^5.13.2"
autoflake = "^2.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
